okx_api/requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.20.0